[RECIPE-PROJECT] Done login form and added Auth context
