+++ recipe-hooks-project/src/App.test.js
@@ -1,9 +1,18 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import App from './App';
+import React from "react";
+import {render} from "@testing-library/react";
+import App from "./App";
+import {AuthContext} from "./contexts/useAuth";
 
-test('renders learn react link', () => {
-  const { getByText } = render(<App />);
-  const appTitle = getByText(/Recipe App/i);
-  expect(appTitle).toBeInTheDocument();
+test("when not authenticated, renders login", () => {
+  const {container} = render(<App/>);
+  expect(container.querySelector("h3")).toHaveTextContent("Login");
+});
+
+test("when authenticated, do not render login", () => {
+  const {container} = render(
+    <AuthContext.Provider value={true}>
+      <App/>
+    </AuthContext.Provider>
+  );
+  expect(container.querySelector("input[name=email]")).toBeNull();
 });
