+++ recipe-hooks-project/src/contexts/useAuth.js
@@ -1,0 +1,7 @@
+import {createContext, useContext} from "react";
+
+export const AuthContext = createContext();
+
+export function useAuth() {
+  return useContext(AuthContext);
+}
